Fixed DataGeneration not evaluating correctly functions
Before this fix, the evaluation of some functions were not working as intended (i.e. runif(1, -2.4, -2.0) was returning 1.1!). This was due to an incorrect evaluation of the splat function within the dataframe generation.
Now, the definition of the functions uses a named variable ("arg"), where the named variables are placed inside. Additionally, mu->mean... mb.